_tissue: adding gmm cleanup FUN & dynamic matrix check
the cleanup function is using a BayesianMixtureModel and can be used for cleaning the reconstruction results to get a more realistic reconstruction.

The dynamic-matrix check fixes the bug that when the input-matrix is in sparse format building the sdge can lead to wrong dimensions. The check now corrects the matrix format if the given matrix is in sparse.